Consider also sendonly as "active" for timeout
The logic in call_timer_iterator assumes that sendonly is hold, and then
it uses silent_timeout (1h) instead of timeout (1m).

In drachtio-siprec, the SEND channel actually receives media, so we need
to use the same timeout of 1m. We also don't have real sendonly channels
so only if both SEND and RECV are unset, silent_timeout should be used.